hedonic model using statsmodels.api
Hedonic Regression by a single quarter or MULTIPLE. We isolated the variable we had wanted to explore, which had been the log of the price per sqft value, which is saved as a float initially. Upon creating the quarterly based, analysis, we are now able to simplify the housing price assessment.